+++ frontend/src/features/store/genresReducer/reducer.js
@@ -21,12 +21,48 @@ const loadGenres = createAsyncThunk(
   },
 );
 
+const loadUserGenres = createAsyncThunk(
+  'genres/initGenres',
+  async () => {
+    const response = await fetch('/api/userGenre');
+    const data = await response.json();
+
+    if (data.error) {
+      console.log(data.error);
+      throw data.error;
+    }
+    return data;
+  },
+);
+
 const addGenre = createAsyncThunk(
   'genres/addedUserGenre',
   async (payload) => {
     console.log(payload, ',++++ thunk');
     const response = await fetch('/api/favorite', {
       method: 'POST',
+      headers: { 'content-type': 'application/json' },
+      body: JSON.stringify({
+        genres: payload,
+      }),
+    });
+
+    const data = await response.json();
+
+    console.log('data', data);
+    if (data.error) {
+      throw data.error;
+    }
+    return data;
+  },
+);
+
+const editGenre = createAsyncThunk(
+  'genres/addedUserGenre',
+  async (payload) => {
+    console.log(payload, ',++++ thunk');
+    const response = await fetch('/api/favorite', {
+      method: 'PUT',
       headers: { 'content-type': 'application/json' },
       body: JSON.stringify({
         genres: payload,
@@ -63,6 +99,15 @@ const genresSlice = createSlice({
         const newGenres = action.payload;
         state.genres = newGenres;
       })
+      .addCase(loadUserGenres.rejected, (state, action) => {
+        // Сценарий провала — загрузка не увенчалась успехом
+        state.error = action.error.message;
+      })
+      .addCase(loadUserGenres.fulfilled, (state, action) => {
+        // Успешный случай — загрузка прошла хорошо
+        const newGenres = action.payload;
+        state.userGenre = newGenres;
+      })
       .addCase(addGenre.rejected, (state, action) => {
         state.error = action.error.message;
       })
@@ -71,6 +116,15 @@ const genresSlice = createSlice({
           user_id: action.payload.user_id,
           genre: action.payload.resultUser,
         };
+      })
+      .addCase(editGenre.rejected, (state, action) => {
+        state.error = action.error.message;
+      })
+      .addCase(editGenre.fulfilled, (state, action) => {
+        state.userGenre = {
+          user_id: action.payload.user_id,
+          genre: action.payload.editGenre,
+        };
       });
   },
 });
@@ -78,6 +132,8 @@ const genresSlice = createSlice({
 export {
   loadGenres,
   addGenre,
+  loadUserGenres,
+  editGenre,
 };
 
 // export const {
