+++ frontend/src/features/store/profileReducer/reducer.js
@@ -1,0 +1,5 @@
+import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
+
+const initialState = {
+  
+}
